Made APIKey part of the sign function.
The goal of this is to only add the APIKey at the last minute, when actually
creating the request. This way, the `show` implementation on queries won't
display it. It also avoids adding it when signing the method, since that's not
required.

The only downside of this approach is the horrible "append" helper function.