Added PNGImage type and refactored network code.
The problem the PNG image posed was that the code relied on the assumption that
the result of a network call would always be JSON text. It turns out that it's
not the case for the PNG images that result from a getAvatar call, which are
either a normal JSON envelope containing an URL to the avatar, or directly a PNG
image.

This led to this refactor, which introduces a more generic Decode class, that
allows a type to specify how to construct a value from a Network response; most
types will default to just silently expecting JSON, as it was before, while the
PNGImage type can do a proper check of the content type.