Removed MonadSign and MonadNetwork, added Config.
Instead of having assumptions on the base class of the stack (how does one sign
a request in IO? In Identity?), I let the user provide a simple Config that
exposes all the needed functions. The config needs to know in which base monad
it is, which in turn requires some liftBase magic.